requirements.txt
# Core dependencies for Gut Health RAG Pipeline
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-core>=0.1.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Additional utilities
pathlib2>=2.3.7
typing-extensions>=4.5.0

# Optional: GPU support (uncomment if you have CUDA)
# torch>=2.0.0
# faiss-gpu>=1.7.4